Reimplemented `get_invalid_moves` using JAX vmap.
Moved the internal function `_maybe_set_invalid_move` as a public global function and renamed it to `get_action_is_invalid`. The function also takes in an additional argument `my_killed_pieces`. The previous form of this function used this variable as a non-local variable. This change is an improvement because it makes the function pure.